refactor(harness): remove the 'veadk agentkit harness' command group
Harness commands now live solely under the top-level `veadk harness` group.
`veadk agentkit harness invoke` was a leftover thin HTTP client fully superseded
by `veadk harness invoke` (which also accepts explicit --url/--key), and its
`add` was already removed. Drop the whole group; move the shared `_harness_request`
helper into cli_harness.py (its only remaining caller); delete the group's
contract tests; update the runtime docs that referenced it.
