+++ docs/content/docs/framework/agent/runtime.en.mdx
@@ -88,7 +88,7 @@ agent = Agent(
 
 The deployed [Harness server](/en/docs/cli/harness-cli) exposes the same switch:
 the harness spec's `runtime` field, and the `--runtime` option on
-`veadk agentkit harness add/invoke` (`adk` or `codex`).
+`veadk harness add/invoke` (`adk` or `codex`).
 
 ## Adding a custom runtime
