+++ docs/content/docs/framework/agent/runtime.mdx
@@ -70,7 +70,7 @@ agent = Agent(
 
 ## 在 Harness Server 中使用
 
-部署的 [Harness server](/cn/docs/cli/harness-cli) 也暴露了运行时开关：harness 规格的 `runtime` 字段，以及 `veadk agentkit harness add/invoke` 的 `--runtime` 选项（`adk` 或 `codex`）。
+部署的 [Harness server](/cn/docs/cli/harness-cli) 也暴露了运行时开关：harness 规格的 `runtime` 字段，以及 `veadk harness add/invoke` 的 `--runtime` 选项（`adk` 或 `codex`）。
 
 ## 扩展自定义运行时
